Move etcd metrics to separate package
Allows importing pkg/metrics without pulling in pkg/etcd, which was causing an import loop in a follow-up commit.
