+++ pkg/etcd/snapshot.go
@@ -23,6 +23,7 @@ import (
 	"github.com/k3s-io/k3s/pkg/daemons/config"
 	"github.com/k3s-io/k3s/pkg/etcd/s3"
 	"github.com/k3s-io/k3s/pkg/etcd/snapshot"
+	"github.com/k3s-io/k3s/pkg/etcd/snapshotmetrics"
 	"github.com/k3s-io/k3s/pkg/util"
 	"github.com/k3s-io/k3s/pkg/util/metrics"
 	"github.com/k3s-io/k3s/pkg/version"
@@ -205,7 +206,7 @@ func (e *ETCD) Snapshot(ctx context.Context) (*managed.SnapshotResult, error) {
 // metrics do not overlap.
 func (e *ETCD) snapshot(ctx context.Context) (_ *managed.SnapshotResult, rerr error) {
 	snapshotStart := time.Now()
-	defer metrics.ObserveWithStatus(snapshotSaveCount, snapshotStart, rerr)
+	defer metrics.ObserveWithStatus(snapshotmetrics.SaveCount, snapshotStart, rerr)
 
 	if !e.snapshotMu.TryLock() {
 		return nil, errors.New("snapshot save already in progress")
@@ -261,7 +262,7 @@ func (e *ETCD) snapshot(ctx context.Context) (_ *managed.SnapshotResult, rerr er
 
 	saveStart := time.Now()
 	_, err = snapshotv3.SaveWithVersion(ctx, e.client.GetLogger(), *cfg, snapshotPath)
-	metrics.ObserveWithStatus(snapshotSaveLocalCount, saveStart, err)
+	metrics.ObserveWithStatus(snapshotmetrics.SaveLocalCount, saveStart, err)
 
 	if err != nil {
 		sf = &snapshot.File{
@@ -343,7 +344,7 @@ func (e *ETCD) snapshot(ctx context.Context) (_ *managed.SnapshotResult, rerr er
 			if s3client, err := e.getS3Client(ctx); err != nil {
 				logrus.Warnf("Unable to initialize S3 client: %v", err)
 				if !errors.Is(err, s3.ErrNoConfigSecret) {
-					metrics.ObserveWithStatus(snapshotSaveS3Count, s3Start, err)
+					metrics.ObserveWithStatus(snapshotmetrics.SaveS3Count, s3Start, err)
 					err = pkgerrors.WithMessage(err, "failed to initialize S3 client")
 					sf = &snapshot.File{
 						Name:     f.Name(),
@@ -363,7 +364,7 @@ func (e *ETCD) snapshot(ctx context.Context) (_ *managed.SnapshotResult, rerr er
 				// upload will return a snapshot.File even on error - if there was an
 				// error, it will be reflected in the status and message.
 				sf, err = s3client.Upload(ctx, snapshotPath, extraMetadata, now)
-				metrics.ObserveWithStatus(snapshotSaveS3Count, s3Start, err)
+				metrics.ObserveWithStatus(snapshotmetrics.SaveS3Count, s3Start, err)
 				if err != nil {
 					logrus.Errorf("Error received during snapshot upload to S3: %s", err)
 				} else {
@@ -684,7 +685,7 @@ func (e *ETCD) ReconcileSnapshotData(ctx context.Context) error {
 // the provided SnapshotResult are deleted, even if they are within a retention window.
 func (e *ETCD) reconcileSnapshotData(ctx context.Context, res *managed.SnapshotResult) (rerr error) {
 	reconcileStart := time.Now()
-	defer metrics.ObserveWithStatus(snapshotReconcileCount, reconcileStart, rerr)
+	defer metrics.ObserveWithStatus(snapshotmetrics.ReconcileCount, reconcileStart, rerr)
 
 	// make sure the core.Factory is initialized. There can
 	// be a race between this core code startup.
@@ -698,7 +699,7 @@ func (e *ETCD) reconcileSnapshotData(ctx context.Context, res *managed.SnapshotR
 	// Get snapshots from local filesystem
 	localStart := time.Now()
 	snapshotFiles, err := e.listLocalSnapshots()
-	metrics.ObserveWithStatus(snapshotReconcileLocalCount, localStart, err)
+	metrics.ObserveWithStatus(snapshotmetrics.ReconcileLocalCount, localStart, err)
 	if err != nil {
 		return err
 	}
@@ -711,12 +712,12 @@ func (e *ETCD) reconcileSnapshotData(ctx context.Context, res *managed.SnapshotR
 		if s3client, err := e.getS3Client(ctx); err != nil {
 			logrus.Warnf("Unable to initialize S3 client: %v", err)
 			if !errors.Is(err, s3.ErrNoConfigSecret) {
-				metrics.ObserveWithStatus(snapshotReconcileS3Count, s3Start, err)
+				metrics.ObserveWithStatus(snapshotmetrics.ReconcileS3Count, s3Start, err)
 				return pkgerrors.WithMessage(err, "failed to initialize S3 client")
 			}
 		} else {
 			s3Snapshots, err := s3client.ListSnapshots(ctx)
-			metrics.ObserveWithStatus(snapshotReconcileS3Count, s3Start, err)
+			metrics.ObserveWithStatus(snapshotmetrics.ReconcileS3Count, s3Start, err)
 			if err != nil {
 				logrus.Errorf("Error retrieving S3 snapshots for reconciliation: %v", err)
 			} else {
