+++ pkg/metrics/metrics.go
@@ -8,7 +8,7 @@ import (
 	"github.com/k3s-io/k3s/pkg/agent/https"
 	"github.com/k3s-io/k3s/pkg/agent/loadbalancer"
 	"github.com/k3s-io/k3s/pkg/daemons/config"
-	"github.com/k3s-io/k3s/pkg/etcd"
+	"github.com/k3s-io/k3s/pkg/etcd/snapshotmetrics"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	lassometrics "github.com/rancher/lasso/pkg/metrics"
 	rdmetrics "github.com/rancher/remotedialer/metrics"
@@ -38,7 +38,7 @@ func init() {
 	// same for loadbalancer metrics
 	loadbalancer.MustRegister(DefaultRegisterer)
 	// and etcd snapshot metrics
-	etcd.MustRegister(DefaultRegisterer)
+	snapshotmetrics.MustRegister(DefaultRegisterer)
 	// and remotedialer metrics
 	rdmetrics.MustRegister(DefaultRegisterer)
 }
